Add a quick-and-dirty i-regex parser
Completely untested but this should be a complete implementation of the
syntax described in https://www.rfc-editor.org/rfc/rfc9485.pdf.

I'm sure there's going to be a number of bugs here. I hope there's a
large library of tests somewhere to validate everything (especially the
error checking)... I'm hoping that there's something in the ECMAScript
test suite.